[finance_report_vision] refactor(coverage): fold check_source_coverage_matrix into common/coverage
Smallest and last slice of the common/ssot/ dissolution series (see #1564,
#1565, #1566, #1567): this single-file gate validates the source-coverage
matrix's proof-level trust contract, which is squarely coverage's own domain
-- it belongs alongside coverage's other policy/lcov machinery, not in the
ssot junk drawer.

- git mv common/ssot/check_source_coverage_matrix.py ->
  common/coverage/check_source_coverage_matrix.py, preserving history.
- Fix its two consumers: the tools/*.py one-line wrapper and
  tests/tooling/test_source_coverage_matrix.py (both the dotted-module and
  from-import styles).
- No parents[N] fix needed (common/ssot/ and common/coverage/ are the same
  nesting depth). No doc/YAML path-string references existed for this file.

Verified: check_package_contract, check_authority_reconcile (coverage still
CODE-ONLY, llm=0), check_ac_index, lint_doc_consistency, check_manifest,
check_workflow_contract, full tests/tooling/ suite (1843 passed, 1 skipped),
and ruff check on both changed files all pass clean.
