+++ tools/check_source_coverage_matrix.py
@@ -10,7 +10,7 @@ ROOT_DIR = Path(__file__).resolve().parents[1]
 if str(ROOT_DIR) not in sys.path:
     sys.path.insert(0, str(ROOT_DIR))
 
-from common.ssot.check_source_coverage_matrix import main  # noqa: E402
+from common.coverage.check_source_coverage_matrix import main  # noqa: E402
 
 if __name__ == "__main__":  # pragma: no cover
     raise SystemExit(main())
